FilterCaption: user-titled captions for editable_by_user / needs_naming
Two User-typed query attrs were missing from PARAM_LOOKUPS, so
their captions fell through to the raw-value path and printed
the user id ("editable_by_user: 1") instead of the title
("editable_by_user: Nathan Wilson (nathan)"). Adds both to
PARAM_LOOKUPS + the missing `query_editable_by_user` i18n.

Two new tests in filter_caption_test pin the behavior.

Also collapses `Header::FiltersHelper` to its remaining public
API (`add_query_filters` — the view→Phlex bridge):
- inlines `type_tags_to_label` + `SENTINEL_TYPE_TAGS` as a
  private method + constant on `FilterCaption`
- drops the `register_value_helper :type_tags_to_label` line
- deletes `test/helpers/header/filters_helper_test.rb` (its 6
  tests covered only `type_tags_to_label`, already exercised
  indirectly by `test_type_tags_param_localizes_via_helper` on
  the Phlex side).

Three other now-orphan helper tests (`confidence_val_as_label`
× 2, `filter_truncate_joined_string` × 1) were already errored
in CI because the methods moved to FilterCaption on this branch.
