+++ test/views/controllers/application/content/header/index_bar/filter_caption_test.rb
@@ -99,6 +99,31 @@ module Views::Controllers::Application::Content
       assert_no_html(html, "#caption-truncated .small b i")
     end
 
+    # Both `editable_by_user` (SpeciesLists) and `needs_naming`
+    # (Observations) are User-typed query attrs. Before this PR
+    # they weren't in PARAM_LOOKUPS, so the caption fell through
+    # to the raw-value formatter and printed the user id
+    # ("editable_by_user: 1") instead of the user's title.
+    def test_editable_by_user_renders_user_title_not_id
+      user = users(:rolf)
+      query = Query.lookup_and_save(:SpeciesList, editable_by_user: user.id)
+
+      html = render_for(query)
+
+      assert_html(html, "#caption-truncated .small b",
+                  text: user.unique_text_name)
+    end
+
+    def test_needs_naming_renders_user_title_not_id
+      user = users(:rolf)
+      query = Query.lookup_and_save(:Observation, needs_naming: user.id)
+
+      html = render_for(query)
+
+      assert_html(html, "#caption-truncated .small b",
+                  text: user.unique_text_name)
+    end
+
     def test_project_lookup_does_not_italicize
       project = projects(:bolete_project)
       query = Query.lookup_and_save(:Observation, projects: [project.id])
